+++ src/evebox/agent.rs
@@ -4,6 +4,5 @@
 use crate::prelude::*;
 
 pub(crate) fn container_name(context: &Context) -> String {
-    let parent = context.root.file_name().unwrap().to_string_lossy();
-    format!("{}-evectl-evebox-agent", parent)
+    format!("{}-evebox-agent", context.container_prefix())
 }